数据类型更改

+++ 校内比赛/18-19年寒假新乡市友谊赛(普及-)(第二期)/T2 午时已到.cpp
@@ -1,7 +1,7 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-int k,n;
+long long k,n;
 
 void init()
 {
